# server/openapi.yaml
openapi: 3.1.0
info:
  title: Voxels public read API
  version: '1.0.0'
  description: |
    The read-only part of the Voxels API. Everything here is unauthenticated GET.

    Write routes, sign-in, admin, moderation, livekit, radio, metrics, guest
    passes and the internal /grid/* routes are deliberately not described.

    Most handlers answer with an envelope: `{"success": true, "<field>": ...}`,
    where `<field>` is plural for a list and singular for one record. A failed
    lookup answers `{"success": false}`, usually with status 400 and sometimes
    with 200, so check the flag rather than the status.
servers:
  - url: https://www.voxels.com
...
paths:
...
  /api/womps.json:
    get:
      operationId: listWomps
      tags: [womps]
      summary: The newest womps across the world
      description: |
        A womp is a photograph somebody took in world. The feed carries the
        `public` and `broadcast` kinds; `profile` womps, which their author kept
        off the public feed, and `report` womps are not listed here.
      parameters:
        - $ref: '#/components/parameters/WompLimit'
        - name: kind
          in: query
          description: The only accepted value is `broadcast`, which narrows the
            feed to broadcasts. Anything else is ignored and you get both kinds.
          schema:
            type: string
            enum: [broadcast]
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  womps:
                    type: array
                    items:
                      $ref: '#/components/schemas/Womp'
        '404':
          $ref: '#/components/responses/Failure'
  /api/womps/{id}.json:
    get:
      operationId: getWomp
      tags: [womps]
      summary: One womp
      parameters:
        - $ref: '#/components/parameters/WompId'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  womp:
                    $ref: '#/components/schemas/Womp'
        '404':
          $ref: '#/components/responses/Failure'
  /api/womps/{id}.jpg:
    get:
      operationId: getWompImage
      tags: [womps]
      summary: The photograph itself, when it lives in the database
      description: |
        Only womps whose bytes were uploaded to Voxels answer here. The rest are
        hosted elsewhere and only have `image_url`; `image_supplied` on the womp
        record tells you which is which.
      parameters:
        - $ref: '#/components/parameters/WompId'
      responses:
        '200':
          description: ok
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/Failure'
  /api/womps/at/parcel/{parcelId}.json:
    get:
      operationId: listWompsByParcel
      tags: [womps]
      summary: Womps taken on one parcel
      description: Reports are filtered out.
      parameters:
        - name: parcelId
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/WompLimit'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  womps:
                    type: array
                    items:
                      $ref: '#/components/schemas/Womp'
        '404':
          $ref: '#/components/responses/Failure'
  /api/womps/at/space/{spaceId}.json:
    get:
      operationId: listWompsBySpace
      tags: [womps]
      summary: Womps taken in one space
      parameters:
        - name: spaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/WompLimit'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  womps:
                    type: array
                    items:
                      $ref: '#/components/schemas/Womp'
        '404':
          $ref: '#/components/responses/Failure'
  /api/womps/by/{wallet}:
    get:
      operationId: listWompsByAuthor
      tags: [womps]
      summary: Womps one citizen took
      description: |
        Matches `womps.author` exactly, so the wallet has to be cased the way it
        was stored. Reports are filtered out.
      parameters:
        - $ref: '#/components/parameters/Wallet'
        - $ref: '#/components/parameters/WompLimit'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  womps:
                    type: array
                    items:
                      $ref: '#/components/schemas/Womp'
        '404':
          $ref: '#/components/responses/Failure'
...
components:
  parameters:
...
    Wallet:
      name: wallet
      in: path
      required: true
      schema:
        type: string
...
    WompId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    WompLimit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
...
  responses:
    Failure:
      description: The lookup did not land. Some handlers send this with status 200.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Failure'
  schemas:
    Failure:
      type: object
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
    AvatarRef:
      description: |
        A citizen, or the bare lowercased wallet string when no avatar row
        matches. Anywhere an owner or author appears, expect either shape, and in
        the same array: `parcel_users` on a parcel mixes objects and strings.
      oneOf:
        - type: string
        - type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: [string, 'null']
            owner:
              type: string
            created_at:
              type: [string, 'null']
...
    Womp:
      description: A photograph somebody took in world.
      type: object
      properties:
        id:
          type: integer
        author:
          $ref: '#/components/schemas/AvatarRef'
        content:
          type: string
          description: The caption.
        coords:
          type: string
          description: |
            Where the shot was taken, in the same string the play URL uses, for
            example `SW@3558E,2017S,3U`: heading, then east/west, north/south and
            an optional height above ground.
        parcel_id:
          type: [integer, 'null']
        space_id:
          type: [string, 'null']
        parcel_name:
          type: [string, 'null']
        parcel_address:
          type: [string, 'null']
        parcel_island:
          type: [string, 'null']
        space_name:
          type: [string, 'null']
        image_url:
          type: [string, 'null']
        image_supplied:
          type: boolean
          description: |
            True when the bytes are in the database, which is what makes
            `/api/womps/{id}.jpg` work. False means the picture only lives at
            `image_url`. Absent on `/api/womps/{id}.json`.
        created_at:
          type: string
        updated_at:
          type: string
...